c94

+++ common/rust/cctrusted_base/src/api_data.rs
@@ -35,6 +35,6 @@ pub struct CcReport {
  */
 // return structure for get_default_algorithm
 pub struct Algorithm {
-    pub algo_id: u8,
+    pub algo_id: u16,
     pub algo_id_str: String,
 }
